🌱 Split useLocalClusterTools.test.ts into 3 focused test files
Split 1726-line useLocalClusterTools.test.ts into:
- useLocalClusterTools.init.test.ts (~525 lines): initialization, disconnected, demo mode, agent connected
- useLocalClusterTools.crud.test.ts (~815 lines): createCluster, deleteCluster, clusterLifecycle, vCluster ops
- useLocalClusterTools.advanced.test.ts (~590 lines): checkVCluster, refresh, progress, state reset, cleanup

All 94 tests pass (+1 skipped). Zero new lint warnings.
Part of #10439.
