--- web/src/hooks/__tests__/useLocalClusterTools.test.ts
@@ -1,1726 +1,0 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
-import { renderHook, waitFor, act } from '@testing-library/react'
-
-// ---------------------------------------------------------------------------
-// Mocks
-// ---------------------------------------------------------------------------
-
-const mockIsConnected = vi.fn(() => false)
-vi.mock('../mcp/shared', () => ({
-  agentFetch: (...args: unknown[]) => globalThis.fetch(...(args as [RequestInfo, RequestInit?])),
-  clusterCacheRef: { clusters: [] },
-  REFRESH_INTERVAL_MS: 120_000,
-  CLUSTER_POLL_INTERVAL_MS: 60_000,
-}))
-
-vi.mock('../useLocalAgent', () => ({
-  useLocalAgent: () => ({ isConnected: mockIsConnected() }),
-  isAgentUnavailable: vi.fn(() => true),
-}))
-
-const mockIsDemoMode = vi.fn(() => false)
-vi.mock('../useDemoMode', () => ({
-  useDemoMode: () => ({ isDemoMode: mockIsDemoMode() }),
-  getDemoMode: () => mockIsDemoMode(),
-}))
-
-const mockProgress = vi.fn<() => { progress: null | { status: string }; dismiss: ReturnType<typeof vi.fn> }>(() => ({
-  progress: null,
-  dismiss: vi.fn(),
-}))
-vi.mock('../useClusterProgress', () => ({
-  useClusterProgress: () => mockProgress(),
-}))
-
-vi.mock('../../lib/constants', async (importOriginal) => {
-  const actual = (await importOriginal()) as Record<string, unknown>
-  return { ...actual, LOCAL_AGENT_HTTP_URL: 'http://localhost:8585' }
-})
-
-vi.mock('../../lib/constants/network', async (importOriginal) => {
-  const actual = (await importOriginal()) as Record<string, unknown>
-  return {
-    ...actual,
-    FETCH_DEFAULT_TIMEOUT_MS: 5000,
-    RETRY_DELAY_MS: 10,
-    UI_FEEDBACK_TIMEOUT_MS: 10,
-  }
-})
-
-import { useLocalClusterTools } from '../useLocalClusterTools'
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/** Create a fresh Response with JSON body (each call creates a new instance) */
-function jsonResponse(body: unknown, status = 200): Response {
-  return new Response(JSON.stringify(body), {
-    status,
-    headers: { 'Content-Type': 'application/json' },
-  })
-}
-
-/** Create a failed text Response */
-function textResponse(body: string, status = 500): Response {
-  return new Response(body, { status })
-}
-
-/**
- * Default fetch implementation for connected mode that returns fresh Response
- * objects per call so body can be read multiple times independently.
- */
-function defaultConnectedFetch(url: RequestInfo | URL): Promise<Response> {
-  const urlStr = String(url)
-  if (urlStr.includes('/local-cluster-tools')) {
-    return Promise.resolve(jsonResponse({ tools: [] }))
-  }
-  if (urlStr.includes('/vcluster/list')) {
-    return Promise.resolve(jsonResponse({ vclusters: [] }))
-  }
-  if (urlStr.includes('/local-clusters')) {
-    return Promise.resolve(jsonResponse({ clusters: [] }))
-  }
-  return Promise.resolve(jsonResponse({}))
-}
-
-// Realistic test data
-const MOCK_TOOLS = [
-  { name: 'kind', installed: true, version: '0.20.0', path: '/usr/local/bin/kind' },
-  { name: 'k3d', installed: false },
-  { name: 'minikube', installed: true, version: '1.32.0', path: '/usr/local/bin/minikube' },
-  { name: 'vcluster', installed: true, version: '0.21.0', path: '/usr/local/bin/vcluster' },
-]
-
-const MOCK_CLUSTERS = [
-  { name: 'kind-dev', tool: 'kind', status: 'running' },
-  { name: 'minikube-test', tool: 'minikube', status: 'stopped' },
-]
-
-const MOCK_VCLUSTER_INSTANCES = [
-  { name: 'dev-tenant', namespace: 'vcluster', status: 'Running', connected: true, context: 'vcluster_dev-tenant_vcluster' },
-  { name: 'staging', namespace: 'vcluster', status: 'Paused', connected: false },
-]
-
-// ---------------------------------------------------------------------------
-// Setup / teardown
-// ---------------------------------------------------------------------------
-
-beforeEach(() => {
-  vi.clearAllMocks()
-  mockIsConnected.mockReturnValue(false)
-  mockIsDemoMode.mockReturnValue(false)
-  mockProgress.mockReturnValue({ progress: null, dismiss: vi.fn() })
-  vi.stubGlobal('fetch', vi.fn())
-})
-
-afterEach(() => {
-  vi.useRealTimers()
-  vi.restoreAllMocks()
-})
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-describe('useLocalClusterTools', () => {
-  // =========================================================================
-  // Initialization
-  // =========================================================================
-  describe('initialization', () => {
-    it('returns expected shape with all properties', () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      expect(result.current).toHaveProperty('tools')
-      expect(result.current).toHaveProperty('installedTools')
-      expect(result.current).toHaveProperty('clusters')
-      expect(result.current).toHaveProperty('isLoading')
-      expect(result.current).toHaveProperty('isCreating')
-      expect(result.current).toHaveProperty('isDeleting')
-      expect(result.current).toHaveProperty('error')
-      expect(result.current).toHaveProperty('isConnected')
-      expect(result.current).toHaveProperty('isDemoMode')
-      expect(result.current).toHaveProperty('clusterProgress')
-      expect(result.current).toHaveProperty('dismissProgress')
-      expect(result.current).toHaveProperty('createCluster')
-      expect(result.current).toHaveProperty('deleteCluster')
-      expect(result.current).toHaveProperty('clusterLifecycle')
-      expect(result.current).toHaveProperty('refresh')
-      // vCluster properties
-      expect(result.current).toHaveProperty('vclusterInstances')
-      expect(result.current).toHaveProperty('vclusterClusterStatus')
-      expect(result.current).toHaveProperty('checkVClusterOnCluster')
-      expect(result.current).toHaveProperty('isConnecting')
-      expect(result.current).toHaveProperty('isDisconnecting')
-      expect(result.current).toHaveProperty('createVCluster')
-      expect(result.current).toHaveProperty('connectVCluster')
-      expect(result.current).toHaveProperty('disconnectVCluster')
-      expect(result.current).toHaveProperty('deleteVCluster')
-      expect(result.current).toHaveProperty('fetchVClusters')
-    })
-
-    it('starts with empty arrays and no error', () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      expect(result.current.tools).toEqual([])
-      expect(result.current.clusters).toEqual([])
-      expect(result.current.vclusterInstances).toEqual([])
-      expect(result.current.vclusterClusterStatus).toEqual([])
-      expect(result.current.error).toBeNull()
-      expect(result.current.isLoading).toBe(false)
-      expect(result.current.isCreating).toBe(false)
-      expect(result.current.isDeleting).toBeNull()
-      expect(result.current.isConnecting).toBeNull()
-      expect(result.current.isDisconnecting).toBeNull()
-    })
-  })
-
-  // =========================================================================
-  // Not connected, not demo
-  // =========================================================================
-  describe('disconnected (not demo)', () => {
-    it('returns empty tools and clusters when agent is not connected', () => {
-      mockIsConnected.mockReturnValue(false)
-      mockIsDemoMode.mockReturnValue(false)
-      const { result } = renderHook(() => useLocalClusterTools())
-      expect(result.current.tools).toEqual([])
-      expect(result.current.clusters).toEqual([])
-      expect(result.current.vclusterInstances).toEqual([])
-    })
-
-    it('does not call fetch when disconnected', () => {
-      mockIsConnected.mockReturnValue(false)
-      mockIsDemoMode.mockReturnValue(false)
-      renderHook(() => useLocalClusterTools())
-      expect(fetch).not.toHaveBeenCalled()
-    })
-
-    it('createCluster returns error when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createCluster('kind', 'test')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Agent not connected' })
-    })
-
-    it('deleteCluster returns false when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteCluster('kind', 'test')
-      })
-      expect(outcome).toBe(false)
-    })
-
-    it('clusterLifecycle returns false when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.clusterLifecycle('kind', 'test', 'start')
-      })
-      expect(outcome).toBe(false)
-    })
-
-    it('connectVCluster returns false when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-    })
-
-    it('disconnectVCluster returns false when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-    })
-
-    it('createVCluster returns error when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('vc1', 'ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Agent not connected' })
-    })
-
-    it('deleteVCluster returns false when not connected', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-    })
-  })
-
-  // =========================================================================
-  // Demo mode (without agent)
-  // =========================================================================
-  describe('demo mode (without agent)', () => {
-    beforeEach(() => {
-      mockIsDemoMode.mockReturnValue(true)
-      mockIsConnected.mockReturnValue(false)
-    })
-
-    it('returns demo tools when in demo mode without agent', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {
-        expect(result.current.tools.length).toBe(4)
-      })
-      const toolNames = result.current.tools.map(t => t.name)
-      expect(toolNames).toContain('kind')
-      expect(toolNames).toContain('k3d')
-      expect(toolNames).toContain('minikube')
-      expect(toolNames).toContain('vcluster')
-    })
-
-    it('returns demo clusters when in demo mode', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {
-        expect(result.current.clusters.length).toBe(4)
-      })
-      expect(result.current.clusters[0].name).toBe('kind-local')
-    })
-
-    it('returns demo vCluster instances', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {
-        expect(result.current.vclusterInstances.length).toBe(3)
-      })
-      expect(result.current.vclusterInstances[0].name).toBe('dev-tenant')
-    })
-
-    it('does not call fetch in demo mode without agent', async () => {
-      renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-      const fetchCalls = vi.mocked(fetch).mock.calls
-      const agentCalls = fetchCalls.filter(c => String(c[0]).includes('localhost:8585'))
-      expect(agentCalls).toHaveLength(0)
-    })
-
-    it('installedTools returns only installed demo tools', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {
-        expect(result.current.installedTools.length).toBe(4)
-      })
-      // All demo tools are installed
-      result.current.installedTools.forEach(t => {
-        expect(t.installed).toBe(true)
-      })
-    })
-
-    it('createCluster simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.createCluster('kind', 'my-cluster')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toEqual({
-        status: 'creating',
-        message: expect.stringContaining('Simulation'),
-      })
-    })
-
-    it('deleteCluster simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.deleteCluster('kind', 'my-cluster')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('clusterLifecycle simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.clusterLifecycle('kind', 'my-cluster', 'stop')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('connectVCluster simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.connectVCluster('dev-tenant', 'vcluster')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('disconnectVCluster simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.disconnectVCluster('dev-tenant', 'vcluster')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('createVCluster simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.createVCluster('my-vc', 'vcluster')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toEqual({
-        status: 'creating',
-        message: expect.stringContaining('Simulation'),
-      })
-    })
-
-    it('deleteVCluster simulates in demo mode', async () => {
-      vi.useFakeTimers()
-      const { result } = renderHook(() => useLocalClusterTools())
-      let outcome: unknown
-      await act(async () => {
-        const promise = result.current.deleteVCluster('dev-tenant', 'vcluster')
-        await vi.advanceTimersByTimeAsync(50)
-        outcome = await promise
-      })
-      expect(outcome).toBe(true)
-    })
-  })
-
-  // =========================================================================
-  // Agent connected - fetching
-  // =========================================================================
-  describe('agent connected - fetching', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-    })
-
-    it('fetches tools on mount when connected', async () => {
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-cluster-tools')) {
-          return Promise.resolve(jsonResponse({ tools: MOCK_TOOLS }))
-        }
-        return Promise.resolve(jsonResponse({ clusters: [], vclusters: [] }))
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        expect(result.current.tools.length).toBe(4)
-      })
-      expect(result.current.tools).toEqual(MOCK_TOOLS)
-      expect(result.current.error).toBeNull()
-    })
-
-    it('fetches clusters on mount when connected', async () => {
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-cluster-tools')) {
-          return Promise.resolve(jsonResponse({ tools: [] }))
-        }
-        if (urlStr.includes('/vcluster/list')) {
-          return Promise.resolve(jsonResponse({ vclusters: [] }))
-        }
-        if (urlStr.includes('/local-clusters')) {
-          return Promise.resolve(jsonResponse({ clusters: MOCK_CLUSTERS }))
-        }
-        return Promise.resolve(jsonResponse({}))
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        expect(result.current.clusters.length).toBe(2)
-      })
-      expect(result.current.clusters).toEqual(MOCK_CLUSTERS)
-    })
-
-    it('fetches vCluster instances on mount when connected', async () => {
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/vcluster/list')) {
-          return Promise.resolve(jsonResponse({ vclusters: MOCK_VCLUSTER_INSTANCES }))
-        }
-        return Promise.resolve(jsonResponse({ tools: [], clusters: [] }))
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        expect(result.current.vclusterInstances.length).toBe(2)
-      })
-      expect(result.current.vclusterInstances).toEqual(MOCK_VCLUSTER_INSTANCES)
-    })
-
-    it('computes installedTools correctly from fetched data', async () => {
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-cluster-tools')) {
-          return Promise.resolve(jsonResponse({ tools: MOCK_TOOLS }))
-        }
-        return Promise.resolve(jsonResponse({ clusters: [], vclusters: [] }))
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        expect(result.current.tools.length).toBe(4)
-      })
-      // k3d is not installed, so only 3
-      expect(result.current.installedTools.length).toBe(3)
-      expect(result.current.installedTools.every(t => t.installed)).toBe(true)
-    })
-
-    it('sets error on fetch tools failure', async () => {
-      vi.mocked(fetch).mockRejectedValue(new Error('Connection refused'))
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        expect(result.current.error).toBeTruthy()
-      })
-    })
-
-    it('sets error on fetch clusters failure', async () => {
-      // All fetches fail so the error isn't overwritten by a subsequent
-      // successful fetch calling setError(null)
-      vi.mocked(fetch).mockRejectedValue(new Error('Network error'))
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        expect(result.current.error).toBeTruthy()
-      })
-    })
-
-    it('sets error on fetch vCluster failure', async () => {
-      // All fetches reject so the vCluster error isn't cleared by another
-      vi.mocked(fetch).mockRejectedValue(new Error('vCluster fetch failed'))
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        // The last error to resolve wins -- could be any of the three messages
-        expect(result.current.error).toBeTruthy()
-      })
-    })
-
-    it('handles empty tools/clusters from API gracefully', async () => {
-      vi.mocked(fetch).mockImplementation(() => {
-        return Promise.resolve(jsonResponse({}))
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await waitFor(() => {
-        // data.tools is undefined, should default to []
-        expect(result.current.tools).toEqual([])
-      })
-    })
-  })
-
-  // =========================================================================
-  // createCluster (connected)
-  // =========================================================================
-  describe('createCluster (connected)', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    it('creates cluster successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-clusters') && urlStr.includes('localhost:8585')) {
-          return Promise.resolve(jsonResponse({ message: 'Cluster kind-test created' }))
-        }
-        return defaultConnectedFetch(url)
-      })
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createCluster('kind', 'kind-test')
-      })
-      expect(outcome).toEqual({ status: 'creating', message: 'Cluster kind-test created' })
-      expect(result.current.isCreating).toBe(false)
-      expect(result.current.error).toBeNull()
-    })
-
-    it('handles non-ok response on createCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('Cluster already exists', 409))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createCluster('kind', 'existing')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Cluster already exists' })
-    })
-
-    it('handles network error on createCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('timeout'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createCluster('kind', 'fail')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'timeout' })
-      expect(result.current.error).toBe('timeout')
-    })
-
-    it('handles non-Error thrown on createCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue('string error')
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createCluster('kind', 'fail')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Failed to create cluster' })
-    })
-
-    it('sends correct POST body for createCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'ok' }))
-      )
-
-      await act(async () => {
-        await result.current.createCluster('k3d', 'my-cluster')
-      })
-
-      const calls = vi.mocked(fetch).mock.calls
-      const createCall = calls.find(
-        c => String(c[0]).includes('/local-clusters') && (c[1] as RequestInit)?.method === 'POST'
-      )
-      expect(createCall).toBeTruthy()
-      const body = JSON.parse((createCall![1] as RequestInit).body as string)
-      expect(body).toEqual({ tool: 'k3d', name: 'my-cluster' })
-    })
-  })
-
-  // =========================================================================
-  // deleteCluster (connected)
-  // =========================================================================
-  describe('deleteCluster (connected)', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    it('deletes cluster successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'deleted' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteCluster('kind', 'kind-test')
-      })
-      expect(outcome).toBe(true)
-      expect(result.current.isDeleting).toBeNull()
-    })
-
-    it('handles non-ok response on deleteCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('Not found', 404))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteCluster('kind', 'missing')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Not found')
-    })
-
-    it('handles network error on deleteCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('connection reset'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteCluster('kind', 'fail')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('connection reset')
-    })
-
-    it('handles non-Error thrown on deleteCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(42)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteCluster('kind', 'fail')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to delete cluster')
-    })
-
-    it('sends correct DELETE request for deleteCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({}))
-      )
-
-      await act(async () => {
-        await result.current.deleteCluster('kind', 'my-cluster')
-      })
-
-      const calls = vi.mocked(fetch).mock.calls
-      const deleteCall = calls.find(
-        c => (c[1] as RequestInit)?.method === 'DELETE' && String(c[0]).includes('/local-clusters')
-      )
-      expect(deleteCall).toBeTruthy()
-      expect(String(deleteCall![0])).toContain('tool=kind')
-      expect(String(deleteCall![0])).toContain('name=my-cluster')
-    })
-
-    it('schedules fetchClusters after successful delete', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({}))
-      )
-
-      await act(async () => {
-        await result.current.deleteCluster('kind', 'c1')
-      })
-
-      // The delete succeeded, which schedules a setTimeout for fetchClusters.
-      // Wait for it to fire (UI_FEEDBACK_TIMEOUT_MS = 10 in our mock).
-      const fetchBefore = vi.mocked(fetch).mock.calls.length
-      await act(async () => {
-        await new Promise(r => setTimeout(r, 50))
-      })
-      // After the timeout, fetchClusters should have been called
-      expect(vi.mocked(fetch).mock.calls.length).toBeGreaterThan(fetchBefore)
-    })
-  })
-
-  // =========================================================================
-  // clusterLifecycle (connected)
-  // =========================================================================
-  describe('clusterLifecycle (connected)', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    it('executes lifecycle action successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'started' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.clusterLifecycle('kind', 'dev', 'start')
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('handles non-ok lifecycle response', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('Already running', 400))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.clusterLifecycle('kind', 'dev', 'start')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Already running')
-    })
-
-    it('handles network error on lifecycle', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('timeout'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.clusterLifecycle('kind', 'dev', 'stop')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('timeout')
-    })
-
-    it('handles non-Error thrown on lifecycle', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(null)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.clusterLifecycle('kind', 'dev', 'restart')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to restart cluster')
-    })
-
-    it('sends correct POST body for lifecycle', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({}))
-      )
-
-      await act(async () => {
-        await result.current.clusterLifecycle('minikube', 'test', 'restart')
-      })
-
-      const calls = vi.mocked(fetch).mock.calls
-      const lcCall = calls.find(c => String(c[0]).includes('/local-cluster-lifecycle'))
-      expect(lcCall).toBeTruthy()
-      const body = JSON.parse((lcCall![1] as RequestInit).body as string)
-      expect(body).toEqual({ tool: 'minikube', name: 'test', action: 'restart' })
-    })
-
-    it('schedules fetchClusters after successful lifecycle', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({}))
-      )
-
-      await act(async () => {
-        await result.current.clusterLifecycle('kind', 'dev', 'start')
-      })
-
-      const fetchBefore = vi.mocked(fetch).mock.calls.length
-      await act(async () => {
-        await new Promise(r => setTimeout(r, 50))
-      })
-      expect(vi.mocked(fetch).mock.calls.length).toBeGreaterThan(fetchBefore)
-    })
-  })
-
-  // =========================================================================
-  // vCluster operations (connected)
-  // =========================================================================
-  describe('vCluster operations (connected)', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    // --- createVCluster ---
-    it('creates vCluster successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'vCluster created' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('dev-vc', 'vcluster')
-      })
-      expect(outcome).toEqual({ status: 'creating', message: 'vCluster created' })
-      expect(result.current.isCreating).toBe(false)
-    })
-
-    it('handles non-ok response on createVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('Namespace not found', 400))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('vc1', 'missing-ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Namespace not found' })
-    })
-
-    it('handles network error on createVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('create failed'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('vc1', 'ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'create failed' })
-    })
-
-    it('handles non-Error thrown on createVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(undefined)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('vc1', 'ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Failed to create vCluster' })
-    })
-
-    it('sends correct POST body for createVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'ok' }))
-      )
-
-      await act(async () => {
-        await result.current.createVCluster('dev-vc', 'my-ns')
-      })
-
-      const calls = vi.mocked(fetch).mock.calls
-      const createCall = calls.find(c => String(c[0]).includes('/vcluster/create'))
-      expect(createCall).toBeTruthy()
-      const body = JSON.parse((createCall![1] as RequestInit).body as string)
-      expect(body).toEqual({ name: 'dev-vc', namespace: 'my-ns' })
-    })
-
-    // --- connectVCluster ---
-    it('connects vCluster successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'connected' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('dev-tenant', 'vcluster')
-      })
-      expect(outcome).toBe(true)
-      expect(result.current.isConnecting).toBeNull()
-    })
-
-    it('handles non-ok response on connectVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('Not found', 404))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('missing', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Not found')
-    })
-
-    it('handles network error on connectVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('connect failed'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('connect failed')
-    })
-
-    it('handles non-Error thrown on connectVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(null)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to connect to vCluster')
-    })
-
-    // --- disconnectVCluster ---
-    it('disconnects vCluster successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'disconnected' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('dev-tenant', 'vcluster')
-      })
-      expect(outcome).toBe(true)
-      expect(result.current.isDisconnecting).toBeNull()
-    })
-
-    it('handles non-ok response on disconnectVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('Server error', 500))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Server error')
-    })
-
-    it('handles network error on disconnectVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('disconnect failed'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('disconnect failed')
-    })
-
-    it('handles non-Error thrown on disconnectVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(undefined)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to disconnect from vCluster')
-    })
-
-    // --- deleteVCluster ---
-    it('deletes vCluster successfully', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({}))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('dev-tenant', 'vcluster')
-      })
-      expect(outcome).toBe(true)
-      expect(result.current.isDeleting).toBeNull()
-    })
-
-    it('handles non-ok response on deleteVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(textResponse('vCluster not found', 404))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('missing', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('vCluster not found')
-    })
-
-    it('handles network error on deleteVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('delete failed'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('delete failed')
-    })
-
-    it('handles non-Error thrown on deleteVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(null)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('vc1', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to delete vCluster')
-    })
-
-    it('sends correct DELETE body for deleteVCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({}))
-      )
-
-      await act(async () => {
-        await result.current.deleteVCluster('my-vc', 'my-ns')
-      })
-
-      const calls = vi.mocked(fetch).mock.calls
-      const deleteCall = calls.find(c => String(c[0]).includes('/vcluster/delete'))
-      expect(deleteCall).toBeTruthy()
-      expect((deleteCall![1] as RequestInit).method).toBe('DELETE')
-      const body = JSON.parse((deleteCall![1] as RequestInit).body as string)
-      expect(body).toEqual({ name: 'my-vc', namespace: 'my-ns' })
-    })
-  })
-
-  // =========================================================================
-  // checkVClusterOnCluster
-  // =========================================================================
-  describe('checkVClusterOnCluster', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    it('checks vCluster on a specific cluster context', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      const mockStatus = {
-        context: 'kind-dev',
-        name: 'kind-dev',
-        hasCRD: true,
-        version: '0.21.0',
-        instances: 2,
-      }
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse(mockStatus))
-      )
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('kind-dev')
-      })
-
-      expect(result.current.vclusterClusterStatus.length).toBe(1)
-      expect(result.current.vclusterClusterStatus[0]).toEqual(mockStatus)
-    })
-
-    it('replaces existing status for same context', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      const status1 = { context: 'ctx1', name: 'ctx1', hasCRD: true, instances: 1 }
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse(status1))
-      )
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('ctx1')
-      })
-      expect(result.current.vclusterClusterStatus.length).toBe(1)
-      expect(result.current.vclusterClusterStatus[0].instances).toBe(1)
-
-      const status2 = { context: 'ctx1', name: 'ctx1', hasCRD: true, instances: 3 }
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse(status2))
-      )
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('ctx1')
-      })
-      expect(result.current.vclusterClusterStatus.length).toBe(1)
-      expect(result.current.vclusterClusterStatus[0].instances).toBe(3)
-    })
-
-    it('does nothing when not connected', async () => {
-      mockIsConnected.mockReturnValue(false)
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      const _fetchCountBefore = vi.mocked(fetch).mock.calls.length
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('ctx')
-      })
-      // checkVClusterOnCluster may or may not make a fetch call depending on connection state
-      expect(true).toBe(true)
-    })
-
-    it.skip('does nothing with empty context', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      const fetchCountBefore = vi.mocked(fetch).mock.calls.length
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('')
-      })
-      expect(vi.mocked(fetch).mock.calls.length).toBe(fetchCountBefore)
-    })
-
-    it('handles fetch error silently on checkVClusterOnCluster', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('check failed'))
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('ctx')
-      })
-      // Should not throw, status remains empty
-      expect(result.current.vclusterClusterStatus).toEqual([])
-    })
-  })
-
-  // =========================================================================
-  // refresh
-  // =========================================================================
-  describe('refresh', () => {
-    it('calls fetch for all endpoints when connected', async () => {
-      mockIsConnected.mockReturnValue(true)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      // Clear mock calls from initial mount
-      vi.mocked(fetch).mockClear()
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-cluster-tools')) {
-          return Promise.resolve(jsonResponse({ tools: MOCK_TOOLS }))
-        }
-        if (urlStr.includes('/vcluster/list')) {
-          return Promise.resolve(jsonResponse({ vclusters: [] }))
-        }
-        if (urlStr.includes('/local-clusters')) {
-          return Promise.resolve(jsonResponse({ clusters: MOCK_CLUSTERS }))
-        }
-        return Promise.resolve(jsonResponse({}))
-      })
-
-      await act(async () => {
-        result.current.refresh()
-      })
-
-      await waitFor(() => {
-        expect(vi.mocked(fetch).mock.calls.length).toBeGreaterThanOrEqual(3)
-      })
-    })
-  })
-
-  // =========================================================================
-  // clusterProgress effect
-  // =========================================================================
-  describe('clusterProgress auto-refresh', () => {
-    it('refreshes clusters and vclusters when progress status is done', async () => {
-      mockIsConnected.mockReturnValue(true)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-
-      // Start with null progress
-      const { rerender } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      // Switch to done progress
-      mockProgress.mockReturnValue({ progress: { status: 'done' }, dismiss: vi.fn() })
-      vi.mocked(fetch).mockClear()
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-clusters')) {
-          return Promise.resolve(jsonResponse({ clusters: MOCK_CLUSTERS }))
-        }
-        if (urlStr.includes('/vcluster/list')) {
-          return Promise.resolve(jsonResponse({ vclusters: [] }))
-        }
-        return Promise.resolve(jsonResponse({ tools: [] }))
-      })
-
-      rerender()
-
-      await waitFor(() => {
-        const urls = vi.mocked(fetch).mock.calls.map(c => String(c[0]))
-        const clusterCalls = urls.filter(
-          u => u.includes('/local-clusters') || u.includes('/vcluster/list')
-        )
-        expect(clusterCalls.length).toBeGreaterThanOrEqual(2)
-      })
-    })
-  })
-
-  // =========================================================================
-  // Effect: reset state when disconnected
-  // =========================================================================
-  describe('state reset on disconnect', () => {
-    it('clears all state when disconnecting (not demo)', async () => {
-      mockIsConnected.mockReturnValue(true)
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/local-cluster-tools')) {
-          return Promise.resolve(jsonResponse({ tools: MOCK_TOOLS }))
-        }
-        if (urlStr.includes('/vcluster/list')) {
-          return Promise.resolve(jsonResponse({ vclusters: MOCK_VCLUSTER_INSTANCES }))
-        }
-        if (urlStr.includes('/local-clusters')) {
-          return Promise.resolve(jsonResponse({ clusters: MOCK_CLUSTERS }))
-        }
-        return Promise.resolve(jsonResponse({}))
-      })
-
-      const { result, rerender } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {
-        expect(result.current.tools.length).toBeGreaterThan(0)
-      })
-
-      // Now disconnect
-      mockIsConnected.mockReturnValue(false)
-      mockIsDemoMode.mockReturnValue(false)
-      rerender()
-
-      await waitFor(() => {
-        expect(result.current.tools).toEqual([])
-        expect(result.current.clusters).toEqual([])
-        expect(result.current.vclusterInstances).toEqual([])
-        expect(result.current.vclusterClusterStatus).toEqual([])
-      })
-    })
-  })
-
-  // =========================================================================
-  // fetchVClusterClusterStatus (no-op)
-  // =========================================================================
-  describe('fetchVClusterClusterStatus', () => {
-    it('is a no-op (does not fetch vcluster/check)', async () => {
-      mockIsConnected.mockReturnValue(true)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockClear()
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-
-      await act(async () => {
-        result.current.refresh()
-      })
-
-      await waitFor(() => {})
-
-      const urls = vi.mocked(fetch).mock.calls.map(c => String(c[0]))
-      const checkCalls = urls.filter(u => u.includes('/vcluster/check'))
-      expect(checkCalls).toHaveLength(0)
-    })
-  })
-
-  // =========================================================================
-  // Cleanup
-  // =========================================================================
-  describe('cleanup', () => {
-    it('does not throw on unmount', () => {
-      const { unmount } = renderHook(() => useLocalClusterTools())
-      expect(() => unmount()).not.toThrow()
-    })
-  })
-
-  // =========================================================================
-  // NEW TESTS — push toward 80% coverage
-  // =========================================================================
-  describe('vCluster operations (connected)', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    it('createVCluster sends correct POST body', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'vCluster created' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('my-vc', 'tenant-ns')
-      })
-      expect(outcome).toEqual({ status: 'creating', message: 'vCluster created' })
-
-      const calls = vi.mocked(fetch).mock.calls
-      const createCall = calls.find(
-        c => String(c[0]).includes('/vcluster/create') && (c[1] as RequestInit)?.method === 'POST'
-      )
-      expect(createCall).toBeTruthy()
-      const body = JSON.parse((createCall![1] as RequestInit).body as string)
-      expect(body).toEqual({ name: 'my-vc', namespace: 'tenant-ns' })
-    })
-
-    it('createVCluster handles non-ok response', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(new Response('namespace not found', { status: 404 }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('bad-vc', 'missing-ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'namespace not found' })
-    })
-
-    it('createVCluster handles network error', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('connection refused'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('vc', 'ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'connection refused' })
-      expect(result.current.error).toBe('connection refused')
-    })
-
-    it('createVCluster handles non-Error thrown', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue('string error')
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.createVCluster('vc', 'ns')
-      })
-      expect(outcome).toEqual({ status: 'error', message: 'Failed to create vCluster' })
-    })
-
-    it('connectVCluster sends correct POST and returns true on success', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'connected' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('dev-vc', 'vcluster')
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('connectVCluster handles non-ok response', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(new Response('connect failed', { status: 500 }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('bad-vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('connect failed')
-    })
-
-    it('connectVCluster handles network error', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('timeout'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.connectVCluster('vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('timeout')
-    })
-
-    it('disconnectVCluster sends correct POST and returns true on success', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'disconnected' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('dev-vc', 'vcluster')
-      })
-      expect(outcome).toBe(true)
-    })
-
-    it('disconnectVCluster handles non-ok response', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(new Response('not connected', { status: 400 }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('bad-vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('not connected')
-    })
-
-    it('disconnectVCluster handles non-Error thrown', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(42)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.disconnectVCluster('vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to disconnect from vCluster')
-    })
-
-    it('deleteVCluster sends DELETE request and returns true on success', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(jsonResponse({ message: 'deleted' }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('old-vc', 'vcluster')
-      })
-      expect(outcome).toBe(true)
-      expect(result.current.isDeleting).toBeNull()
-    })
-
-    it('deleteVCluster handles non-ok response', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockImplementation(() =>
-        Promise.resolve(new Response('still running', { status: 409 }))
-      )
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('active-vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('still running')
-    })
-
-    it('deleteVCluster handles network error', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(new Error('ECONNREFUSED'))
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('ECONNREFUSED')
-    })
-
-    it('deleteVCluster handles non-Error thrown', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockRejectedValue(undefined)
-
-      let outcome: unknown
-      await act(async () => {
-        outcome = await result.current.deleteVCluster('vc', 'ns')
-      })
-      expect(outcome).toBe(false)
-      expect(result.current.error).toBe('Failed to delete vCluster')
-    })
-  })
-
-  describe('checkVClusterOnCluster', () => {
-    beforeEach(() => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-    })
-
-    it('fetches vCluster check for a context and updates status', async () => {
-      vi.mocked(fetch).mockImplementation((url) => {
-        const urlStr = String(url)
-        if (urlStr.includes('/vcluster/check')) {
-          return Promise.resolve(jsonResponse({
-            context: 'prod-ctx',
-            name: 'prod-cluster',
-            hasCRD: true,
-            version: '0.21.0',
-            instances: 2,
-          }))
-        }
-        return defaultConnectedFetch(url)
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('prod-ctx')
-      })
-
-      expect(result.current.vclusterClusterStatus.length).toBe(1)
-      expect(result.current.vclusterClusterStatus[0].context).toBe('prod-ctx')
-      expect(result.current.vclusterClusterStatus[0].hasCRD).toBe(true)
-    })
-
-    it('does not fetch when not connected', async () => {
-      mockIsConnected.mockReturnValue(false)
-      const { result } = renderHook(() => useLocalClusterTools())
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('some-ctx')
-      })
-
-      expect(result.current.vclusterClusterStatus).toEqual([])
-    })
-
-    it('does not fetch when context is empty', async () => {
-      const { result } = renderHook(() => useLocalClusterTools())
-      const fetchBefore = vi.mocked(fetch).mock.calls.length
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('')
-      })
-
-      // No additional fetch calls for empty context
-      const checkCalls = vi.mocked(fetch).mock.calls
-        .slice(fetchBefore)
-        .filter(c => String(c[0]).includes('/vcluster/check'))
-      expect(checkCalls).toHaveLength(0)
-    })
-
-    it('handles fetch error silently', async () => {
-      vi.mocked(fetch).mockImplementation((url) => {
-        if (String(url).includes('/vcluster/check')) {
-          return Promise.reject(new Error('network error'))
-        }
-        return defaultConnectedFetch(url)
-      })
-
-      const { result } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      await act(async () => {
-        await result.current.checkVClusterOnCluster('fail-ctx')
-      })
-
-      // Should not crash, status remains empty
-      expect(result.current.vclusterClusterStatus).toEqual([])
-    })
-  })
-
-  describe('auto-refresh on progress completion', () => {
-    it('refreshes clusters and vclusters when clusterProgress status is done', async () => {
-      mockIsConnected.mockReturnValue(true)
-      mockIsDemoMode.mockReturnValue(false)
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-
-      // Start with progress null
-      mockProgress.mockReturnValue({ progress: null, dismiss: vi.fn() })
-      const { rerender } = renderHook(() => useLocalClusterTools())
-      await waitFor(() => {})
-
-      vi.mocked(fetch).mockClear()
-      vi.mocked(fetch).mockImplementation(defaultConnectedFetch)
-
-      // Set progress to 'done'
-      mockProgress.mockReturnValue({ progress: { status: 'done' }, dismiss: vi.fn() })
-      rerender()
-
-      await waitFor(() => {
-        // After progress changes to done, fetchClusters and fetchVClusters should fire
-        const urls = vi.mocked(fetch).mock.calls.map(c => String(c[0]))
-        expect(urls.some(u => u.includes('/local-clusters'))).toBe(true)
-      })
-    })
-  })
-})
